Native code files grooming
