+++ core/smart/detection/src/main/cpp/detector/detector.hpp
