+++ core/smart/detection/src/main/cpp/detector/images/condition_image.cpp
@@ -15,32 +15,21 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef KLICK_R_DETECTION_IMAGE_HPP
-#define KLICK_R_DETECTION_IMAGE_HPP
+#include <opencv2/imgproc/imgproc.hpp>
+#include "condition_image.hpp"
 
-#include <opencv2/core/types.hpp>
+using namespace smartautoclicker;
 
-#include "../utils/scalable_roi.hpp"
 
-namespace smartautoclicker {
-
-    class DetectionImage {
-
-    protected:
-        /** */
-        ScalableRoi roi;
-
-        /**
-         *
-         * @param fullSizeColor
-         * @param scaleRatio
-         */
-        virtual void onNewImageLoaded(std::unique_ptr<cv::Mat> fullSizeColor, std::unique_ptr<cv::Mat> scaledGray) = 0;
-
-    public:
-        void processFullSizeBitmap(cv::Mat* screenMat, double scaleRatio);
-        ScalableRoi getRoi() const;
-    };
+void ConditionImage::onNewImageLoaded(std::unique_ptr<cv::Mat> fullSizeColor, std::unique_ptr<cv::Mat> scaledGray) {
+    fullSizeColorMean = cv::mean(*fullSizeColor);
+    scaledGrayMat = std::move(scaledGray);
 }
 
-#endif //KLICK_R_DETECTION_IMAGE_HPP
+cv::Scalar ConditionImage::getFullSizeColorMean() const {
+    return fullSizeColorMean;
+}
+
+cv::Mat* ConditionImage::getScaledGrayMat() const {
+    return scaledGrayMat.get();
+}
