+++ core/smart/detection/src/main/cpp/detector/images/condition_image.hpp
