+++ core/smart/detection/src/main/cpp/detector/images/screen_image.cpp
@@ -19,9 +19,7 @@
  */
 
 #include <opencv2/imgproc/imgproc.hpp>
-
 #include "screen_image.hpp"
-#include "../utils/log.h"
 
 using namespace smartautoclicker;
