+++ core/smart/detection/src/main/cpp/detector/images/screen_image.hpp
