+++ core/smart/detection/src/main/cpp/jni/jni_bitmap.cpp
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../utils/log.h"
+#include "../logs/log.h"
 #include "jni.hpp"
 
 using namespace smartautoclicker;
