+++ core/smart/detection/src/main/cpp/logs/log.cpp
