+++ core/smart/detection/src/main/cpp/scaling/scalable_roi.cpp
