+++ core/smart/detection/src/main/cpp/scaling/scaling.hpp
